third_party/WebKit/LayoutTests/editing/deleting/merge_paragraph_from_h6.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() =>assert_selection(
        [
            '<div contenteditable>',
                'foo<h6>|bar</h6>',
            '</div>',
        ].join(''),
        'delete',
        [
            '<div contenteditable>',
                'foo|bar',
            '</div>',
        ].join('')),
        '1 deleting line break before H6');

test(() =>assert_selection(
        [
            '<div contenteditable>',
                '<em>foo</em><h6><font color="red">|bar</font></h6>',
            '</div>',
        ].join(''),
        'delete',
        [
            '<div contenteditable>',
                '<em>foo|</em><span style="color: red;">bar</span>',
            '</div>',
        ].join('')),
        '2 deleting line break before H6+FONT');

test(() =>assert_selection(
        [
            '<div contenteditable>',
                '<em>foo</em><h6><font color="blue" style="color:red">|bar</font></h6>',
            '</div>',
        ].join(''),
        'delete',
        [
            '<div contenteditable>',
                '<em>foo|</em><span style="color: red;">bar</span>',
            '</div>',
        ].join('')),
        '3 deleting line break before H6 with style');
</script>
